+++ app/trpc/test-helper.ts
@@ -1,7 +1,7 @@
 import type { TT } from "../db/drizzle-client.server";
 import { createUserCookie } from "../utils/auth";
-import { trpcApp } from "./app.server";
-import { createTrpcAppContext } from "./context.server";
+import { createTrpcAppContext } from "./context";
+import { trpcApp } from "./server";
 
 export async function testTrpcClient(options?: { user?: TT["users"] }) {
   const req = new Request("/dummy");
